+++ lib/src/handler/formatter/formatter.dart
@@ -1,6 +1,7 @@
 part of '../handler.dart';
 
 /// Abstract base for formatting a [LogEntry] into lines of text.
+//ignore: one_member_abstracts
 abstract class LogFormatter {
   const LogFormatter();